Add model.py
