+++ 2-EDA/extract_data.py
@@ -7,7 +7,7 @@ import os
 from os import listdir, walk
 
 sys.path.append('../')
-from help_printing import *
+from help_printing_mg import *
 
 def file_to_data(data_path):
     file_data = listdir(data_path)
@@ -20,24 +20,57 @@ def file_to_data(data_path):
 def make_type_array(data, file_data, type_name):
     data['subID'] = data['ID'].str.split("_", n = 3, expand = True)[1]
     data['subType'] = data['ID'].str.split("_", n = 3, expand = True)[2]
-    type_filter = data['subType'] == type_name
-    filter_data = np.array(data[type_filter])    
-    #print(filter_data.shape, filter_data)
     
     id_data = []
     label_data = []
     
-    for i in range(len(filter_data)):
-        #if i%10000 == 0: print(i)
-        if filter_data[i][2] in file_data:
-            id_data.append(filter_data[i][2])
-            label_data.append(filter_data[i][1])
-        else:
-            pass
-    id_data = np.array(id_data, dtype = str)
-    label_data = np.array(label_data, dtype = int)
+    if type_name == 'normal':
+        sub_id_list = np.array(data['subID'])
+        # Delete same values 
+        sub_id_list = np.array(list(set(sub_id_list)))
+        
+        for i in range(len(sub_id_list)):
+            this_sub_id = sub_id_list[i]
 
-    return id_data, label_data
+            filtering = data['subID']==this_sub_id
+            this_sub_labels = data[filtering]['Label']
+            zero_count = list(this_sub_labels).count('0')
+            if zero_count == 6:
+                id_data.append(this_sub_id) 
+                label_data.append(0)
+                print(this_sub_id)
+            else: continue   
+                
+    elif type_name == 'all':
+        for i in range(len(data)):            
+            if data['subID'][i] in file_data:
+                tmp = data[data['subID'] == data['subID'][i]]
+                
+                label = []
+                for idx in range(len(tmp)):
+                    tt = np.array(tmp['Label'])
+                    label.append(tt[idx])
+                    
+                if data['subID'][i] not in id_data:
+                    id_data.append(data['subID'][i])
+                    label = np.array(label)
+                    label_data.append(label)
+            print(i)    
+    else:
+        type_filter = (data['subType'] == type_name) & (data['Label'] == '1')
+        filter_data = np.array(data[type_filter])
+        for i in range(len(filter_data)):
+            if filter_data[i][2] in file_data:
+                id_data.append(filter_data[i][2])
+                label_data.append(filter_data[i][1])
+    #print(filter_data.shape, filter_data)
+    id_data = np.array(id_data, dtype = str)
+    label_data = np.array(label_data)
+    
+    id_array = id_data[0:len(id_data) - 2]
+    label_array = id_data[0:len(label_data) - 2]
+    
+    return id_array, label_array
 
 def hist_type_array(label_data, type_name, save_dir):
 	n, bins, patches = plt.hist(label_data, 
@@ -54,7 +87,7 @@ def hist_type_array(label_data, type_name, save_dir):
 
 def main():
 #csv_file = '../1-Dataset/stage_2_train.csv'
-	csv_file = '../1-Dataset/dcm_test.csv'
+	csv_file = '../1-Dataset/stage_2_train.csv'
 	if_not_exit(csv_file)
 	csv_data = pd.read_csv(csv_file,sep = ",", dtype = 'unicode')
