--- 4-Model/README.md
@@ -1,5 +1,0 @@
-4-Model -
-				-
-				-
-
-DL model development...
