+++ 4-Model/efficientnet_model_training.py
@@ -18,6 +18,9 @@ from keras.callbacks import ModelCheckpoint
 from keras.applications import VGG16, VGG19, ResNet50
 
 from keras.callbacks import ReduceLROnPlateau
+from keras.callbacks import EarlyStopping
+from keras.callbacks import ModelCheckpoint
+
 from keras import backend as K
 from keras.applications.resnet50 import ResNet50
 
@@ -41,21 +44,13 @@ from PIL import Image
 import os, glob
 import matplotlib.pyplot as plt
 
-# 카테고리 지정하기
-categories = ["epidural", "intraparenchymal", "intraventricular", "subarachnoid", "subdural"]
-nb_classes = len(categories)
-# 이미지 크기 지정하기
-image_w = 128
-image_h = 128
-# 데이터 열기 
-
 base_path = 'E:\\Dataset\\rsna-intracranial-hemorrhage-detection\\brain_diagnosis\\4-Model'
 dataset_path = os.path.join(base_path + '\\dataset')
 
-X_train = np.load(dataset_path + '\\X_train.npy')
-y_train = np.load(dataset_path + '\\y_train.npy')
-X_test = np.load(dataset_path + '\\X_test.npy')
-y_test = np.load(dataset_path + '\\y_test.npy')
+X_train = np.load(dataset_path + '\\X_train.npy', allow_pickle=True)
+y_train = np.load(dataset_path + '\\y_train.npy', allow_pickle=True)
+X_test = np.load(dataset_path + '\\X_test.npy', allow_pickle=True)
+y_test = np.load(dataset_path + '\\y_test.npy', allow_pickle=True)
 
 
 '''
@@ -66,39 +61,42 @@ conv_base.trainable = True
 model = Sequential()
 model.add(conv_base)
 #model.add(Flatten())
+model.add(Dense(1024, activation = 'relu'))
 model.add(Dense(64, activation = 'relu'))
-model.add(Dense(64, activation = 'relu'))
-model.add(Dropout(0.3))
-model.add(Dense(5, activation = 'sigmoid'))
+model.add(Dropout(0.5))
+model.add(Dense(6, activation = 'sigmoid'))
 
-model.compile(optimizer = optimizers.Adam(lr = 1e-5), loss = 'categorical_crossentropy', metrics = ['accuracy'])
+model.compile(optimizer = optimizers.Adam(lr = 1e-4), loss = 'binary_crossentropy', metrics = ['binary_accuracy'])
 model.summary()
+
 '''
 import efficientnet.keras as efn
 from iterstrat.ml_stratifiers import MultilabelStratifiedShuffleSplit
 
 base_model = efn.EfficientNetB2(weights = 'imagenet', include_top = False, pooling = 'avg', input_shape = (128, 128, 3))
 x = base_model.output
-x = Dropout(0.15)(x)
-output = Dense(5, activation = 'sigmoid')(x)
+x = Dropout(0.5)(x)
+output = Dense(6, activation = 'sigmoid')(x)
 model = Model(base_model.input, output)
 #model = multi_gpu_model(resnet50, gpus = None)
 model.summary()
-model.compile(optimizer=optimizers.RMSprop(lr = 1e-5), loss = 'binary_crossentropy',  metrics=['binary_accuracy'])
+model.compile(optimizer=optimizers.Adam(lr = 1e-4), loss = 'binary_crossentropy',  metrics=['binary_accuracy'])
 
+hdf5_file = "E:\\Dataset\\rsna-intracranial-hemorrhage-detection\\brain_diagnosis\\5-Results\\weight_Classification_200514_Adam.hdf5"
 
-hdf5_file = "E:\\Dataset\\rsna-intracranial-hemorrhage-detection\\brain_diagnosis\\5-Results\\weight_Classification_200428.hdf5"
-
+'''
+#datagenerator
 batch_size = 16
-
-
-epochs = 300
+epochs = 50
 train_datagen = ImageDataGenerator()
 val_datagen = ImageDataGenerator()
 
 train_data_flow = train_datagen.flow(X_train, y_train, batch_size = batch_size)
 val_data_flow = val_datagen.flow(X_test, y_test, batch_size = batch_size)
+'''
 
+early_stopping = EarlyStopping(monitor='val_loss', mode='min', verbose=1, patience=3)
+mc = ModelCheckpoint('best_model_classification_200514.h5', monitor='val_loss', mode='min', save_best_only=True)
 
 if os.path.exists(hdf5_file):
     # 기존에 학습된 모델 불러들이기
@@ -109,18 +107,36 @@ else:
 
     #history = model.fit_generator(train_data_flow, epochs = epochs, steps_per_epoch = 25, verbose = 2, validation_data = val_data_flow, validation_steps = 16, workers = 4, callbacks=[ReduceLROnPlateau(monitor='val_loss', factor=0.2, patience=30, verbose=1, mode='auto', min_lr=1e-10)])
     #callbacks = [checkpoint])    
-    history = model.fit(X_train, y_train, validation_data=(X_test, y_test), epochs=500, batch_size=32, callbacks=[ReduceLROnPlateau(monitor='val_loss', factor=0.2, patience=10, verbose=1, mode='auto', min_lr=1e-8)])
+    history = model.fit(X_train, y_train, epochs=100, batch_size=32, verbose = 2, validation_split = 0.2, callbacks=[early_stopping, mc])
     model.save_weights(hdf5_file)
-    
+    model.save('E:\\Dataset\\rsna-intracranial-hemorrhage-detection\\brain_diagnosis\\5-Results\\model_Classification_200514_Adam.h5')
     
     fig, ax = plt.subplots(2, 2, figsize=(10, 7))
 
     ax[0, 0].set_title('loss')
     ax[0, 0].plot(history.history['loss'], 'r')
     ax[0, 1].set_title('acc')
-    ax[0, 1].plot(history.history['acc'], 'b')
+    ax[0, 1].plot(history.history['binary_accuracy'], 'b')
 
     ax[1, 0].set_title('val_loss')
     ax[1, 0].plot(history.history['val_loss'], 'r--')
     ax[1, 1].set_title('val_acc')
-    ax[1, 1].plot(history.history['val_acc'], 'b--')
+    ax[1, 1].plot(history.history['val_binary_accuracy'], 'b--')
+    
+loss_and_metrics = model.evaluate(X_test, y_test, batch_size=32)
+print('## evaluation loss and_metrics ##')
+print(loss_and_metrics)
+
+# 7. 모델 사용하기
+xhat = X_test[0:1]
+yhat = model.predict(xhat)
+print('## yhat ##')
+print(yhat)  
+    
+    
+    
+
+    
+    
+    
+
